requirements.txt
allure-pytest
pytest
pytest-instafail
pytest-order
pytest-rerunfailures
pytest-sugar
pytest-xdist
python-dateutil
python-dotenv
pytest-repeat
PyYAML
requests
selenium
six
SQLAlchemy
urllib3
validators
webdriver-manager
allure-python-commons